webapi/requirements.txt
aiofiles==0.3.1
aioredis==0.3.3
asyncpg==0.12.0
bcrypt==3.1.3
cffi==1.10.0
hiredis==0.2.0
httptools==0.0.9
pycparser==2.18
PyJWT==1.5.2
pytimeparse==1.1.6
PyYAML==3.12
sanic==0.6.0
six==1.10.0
ujson==1.35
uvloop==0.8.0
websockets==3.3
